+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.1.4] - 2022-01-11
 ### Added
 - Added the OpenTelemetry to k8s manifests that is for the experiment. (#VARLOG-649/#563)
 - Added `--batch-size` option to `varlogcli` to set the size of append batch. (#VARLOG-653/570)
@@ -69,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-12-15
 
 
-[Unreleased]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.3...HEAD
+[Unreleased]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.4...HEAD
+[0.1.4]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.daumkakao.com/varlog/varlog/compare/v0.1.0...v0.1.1